% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leni_sem.R
\name{leni_sem}
\alias{leni_sem}
\title{Generate Linearized Structural Equation Model Syntax for Lavaan}
\usage{
leni_sem(
  target_fx = "cubic",
  theta = NULL,
  start_values = NULL,
  number_obs = 5,
  center_obs = 1,
  spacing = 1,
  y.name = NULL,
  time.name = NULL,
  verbose = FALSE
)
}
\arguments{
\item{target_fx}{The target function. Can either be a predefined function
from the \code{nonlinear_function_library}, or a string defining a custom
function.}

\item{theta}{Vector of parameter names for the target function. Populated
automatically for predefined functions, but required for custom functions.}

\item{start_values}{(Optional) Start values for the means of the nonlinear latent
factors (defined as phantoms). Defaults to 0 for all factor means.}

\item{number_obs}{(Integer) Number of time points. Enforced through rounding.}

\item{center_obs}{(Optional) Time point where time is defined as 0. Defaults
to the first observation.}

\item{spacing}{Amount of time between observations. If a single numeric value
is provided, then defaults to equal spacing. For unequal spacing, a vector
of values representing the spacing between adjacent time points should be
provided. If a vector is provided, the \code{number_obs} argument is ignored.}

\item{y.name}{(Optional) Provide a string indicating a custom name for the
repeated measures outcome.}

\item{time.name}{(Optional) Provide a string indicating a custom name for the
repeated measure of time (e.g., age, wave, etc.). Cannot be a substring of
an element in \code{theta}.}

\item{verbose}{(Boolean) If \code{TRUE}, outputs a reader-friendly version
of generated syntax into the console.}
}
\description{
Read in linear regression or mixed-effects (multilevel) model
output and generate nonlinear parameter estimates (and standard errors)
through a series of transformations. Analytic and bootstrap estimates
possible. Both main effects and interactions can be transformed. Addition
options can be passed to the bootstrap procedure.
}
\examples{
custom_fx <- "th1 - (th1 - th2) * exp(-th3 * (ti - 1))"

syntax <- leni_sem(target_fx = custom_fx, theta = c("th1", "th2", "th3"),
                   number_obs = 5, spacing = 1.5, y.name = "RT", time.name = "ti",
                   verbose = TRUE)

}
